requirements.txt
azure-ai-textanalytics
azure-core